+++ src/ccd/ccd.cpp
@@ -159,12 +159,12 @@ bool point_edge_ccd_3D(
 }
 
 bool point_edge_ccd(
-    const Eigen::VectorXd& p_t0,
-    const Eigen::VectorXd& e0_t0,
-    const Eigen::VectorXd& e1_t0,
-    const Eigen::VectorXd& p_t1,
-    const Eigen::VectorXd& e0_t1,
-    const Eigen::VectorXd& e1_t1,
+    const Eigen::VectorX3d& p_t0,
+    const Eigen::VectorX3d& e0_t0,
+    const Eigen::VectorX3d& e1_t0,
+    const Eigen::VectorX3d& p_t1,
+    const Eigen::VectorX3d& e0_t1,
+    const Eigen::VectorX3d& e1_t1,
     double& toi,
     double conservative_rescaling)
 {