+++ src/ccd/ccd.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <Eigen/Core>
+#include <ipc/utils/eigen_ext.hpp>
 
 namespace ipc {
 
@@ -65,12 +66,12 @@ bool edge_edge_ccd(
 // 2D or 3D
 
 bool point_edge_ccd(
-    const Eigen::VectorXd& p_t0,
-    const Eigen::VectorXd& e0_t0,
-    const Eigen::VectorXd& e1_t0,
-    const Eigen::VectorXd& p_t1,
-    const Eigen::VectorXd& e0_t1,
-    const Eigen::VectorXd& e1_t1,
+    const Eigen::VectorX3d& p_t0,
+    const Eigen::VectorX3d& e0_t0,
+    const Eigen::VectorX3d& e1_t0,
+    const Eigen::VectorX3d& p_t1,
+    const Eigen::VectorX3d& e0_t1,
+    const Eigen::VectorX3d& e1_t1,
     double& toi,
     double conservative_rescaling = 0.8);