+++ src/distance/edge_edge.hpp
@@ -123,7 +123,7 @@ void edge_edge_distance_gradient(
     grad.resize(4 * dim);
     grad.setZero();
 
-    Eigen::VectorXd local_grad;
+    Eigen::VectorX9d local_grad;
     switch (dtype) {
     case EdgeEdgeDistanceType::EA0_EB0:
         point_point_distance_gradient(ea0, eb0, local_grad);
@@ -227,7 +227,7 @@ void edge_edge_distance_hessian(
     hess.resize(4 * dim, 4 * dim);
     hess.setZero();
 
-    Eigen::MatrixXd local_hess;
+    Eigen::MatrixXX9d local_hess;
     switch (dtype) {
     case EdgeEdgeDistanceType::EA0_EB0:
         point_point_distance_hessian(ea0, eb0, local_hess);