+++ CHANGELOG.md
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 #### Fixed
 -->
 
+### 2021-01-09 <!--([XXXXXXX](https://github.com/ipc-sim/ipc-toolkit/commit/XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX))-->
+#### Changed
+* Replaced VectorXd and MatrixXd with static size versions for local gradient and hessians
+
 ### 2020-11-20 ([93143ad](https://github.com/ipc-sim/ipc-toolkit/commit/93143ad9b31030cde7324a83354268021e1cb9da))
 #### Changed
 * Removed TBB parallelization form the hash grid because we get better performance without it.